requirements.txt
dj-database-url==0.5.0
Django==2.1.3
django-debug-toolbar==1.10.1
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
python-decouple==3.1
tox==3.5.3
